fix: Поправил замечания по ревью. Исправил реализации классов *Exception.java Добавил проверки валидации id для сущностей и изменил тип на long.

+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -31,9 +31,9 @@ public class Film extends Entity implements Comparable<Film> {
     private int duration;
 
     @Setter(value = AccessLevel.PRIVATE)
-    private Set<Integer> likes;
+    private Set<Long> likes;
 
-    public Set<Integer> getLikes() {
+    public Set<Long> getLikes() {
         if (likes == null) {
             return likes = new HashSet<>();
         } else return likes;
